version_4p01_with_vision_agent/pyproject.toml
[project]
name = "version-4p01-with-vision-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.4",
    "asyncclick>=8.1.8",
    "click>=8.2.1",
    "google-adk>=1.1.1",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.2",
    "langgraph>=0.4.7",
    "mcp>=1.9.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
